Project update

+++ database/migrations/2025_02_18_064819_create_inventory_items_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->string('item_name');
             $table->integer('quantity');
-            $table->decimal('unit_price', 10, 2);
-            $table->decimal('package_volume', 10, 2);
+            $table->integer('unit_price');
+            $table->bigInteger('package_volume');
             $table->timestamps();
         });
     }
